packages/io-coding-agent/pyproject.toml
[project]
name = "io-coding-agent"
version = "0.1.0"
description = "Main IO coding agent CLI"
requires-python = ">=3.11"
dependencies = [
  "agent-client-protocol>=0.8.1,<1.0",
  "io-ai",
  "io-agent-core",
  "io-tui",
  "httpx>=0.28.1",
  "prompt_toolkit>=3.0.50",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "rich>=13.9.4",
]

[project.scripts]
io = "io_cli.cli:main"
io-acp = "io_cli.acp_adapter.entry:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/io_cli"]
